+++ feature/yesterdaylist/src/main/java/com/poptato/yesterdaylist/YesterdayListScreen.kt
@@ -28,7 +28,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.poptato.component.PoptatoButton
+import com.poptato.ui.common.PoptatoButton
 import com.poptato.design_system.Complete
 import com.poptato.design_system.Gray00
 import com.poptato.design_system.Gray10
